chore: split panel tests into multiple files
